+++ R/utils.R
@@ -151,6 +151,21 @@ parse_only = function(code) {
 #' inherits(z, 'try-error')
 try_silent = function(expr) try(expr, silent = TRUE)
 
+#' Try an expression and see if it throws an error
+#'
+#' Use \code{\link{tryCatch}()} to check if an expression throws an error.
+#' @inheritParams try_silent
+#' @return \code{TRUE} (error) or \code{FALSE} (success).
+#' @export
+#' @examples
+#' xfun::try_error(stop('foo'))  # TRUE
+#' xfun::try_error(1:10)  # FALSE
+try_error = function(expr) {
+  err = FALSE
+  tryCatch(expr, error = function(e) err <<- TRUE)
+  err
+}
+
 #' Retry calling a function for a number of times
 #'
 #' If the function returns an error, retry it for the specified number of